+++ Examples/Event Sourcing/Examples.EventSourcing/Startup.cs
@@ -30,12 +30,12 @@ namespace Examples
                 .AddServerFaults()
                 .AddEventStoreService<CosmosAggregate>()
                 .AddCosmosDbRepository<CosmosAggregate>(_config)
-                .AddEventStoreService<MongoAggregate>()
-                .AddMongoDbRepository<MongoAggregate>(_config)
-                .AddEventStoreService<RavenAggregate>()
-                .AddRavenDbRepository<RavenAggregate>(_config)
-                .AddEventStoreService<EfCoreSqlAggregate>()
-                .AddEfCoreSqlRepository<EfCoreSqlAggregate>(_config)
+                //.AddEventStoreService<MongoAggregate>()
+                //.AddMongoDbRepository<MongoAggregate>(_config)
+                //.AddEventStoreService<RavenAggregate>()
+                //.AddRavenDbRepository<RavenAggregate>(_config)
+                //.AddEventStoreService<EfCoreSqlAggregate>()
+                //.AddEfCoreSqlRepository<EfCoreSqlAggregate>(_config)
                 ;
 
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);